pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "meterocr"
version = "0.1.0"
description = "Water meter OCR using HOG + scikit-learn"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8",
    "numpy>=1.26",
    "scikit-image>=0.22",
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "pandas>=2.1",
    "pyyaml>=6.0",
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
meterocr = "meterocr.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
]
